// src/client/models/coverage.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Navitia API - SNCF
 * Complete API documentation for Navitia/SNCF public transportation API. Based on https://doc.navitia.io
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface Coverage {
    /**
     * Coverage area ID
     */
    'id'?: string;
    /**
     * Start production date (YYYYMMDD)
     */
    'start_production_date'?: string;
    /**
     * End production date (YYYYMMDD)
     */
    'end_production_date'?: string;
    /**
     * Coverage status
     */
    'status'?: CoverageStatusEnum;
    /**
     * Geographic shape (POLYGON)
     */
    'shape'?: string;
}

export const CoverageStatusEnum = {
    Running: 'running',
    Closed: 'closed'
} as const;

export type CoverageStatusEnum = typeof CoverageStatusEnum[keyof typeof CoverageStatusEnum];